+++ Source/SwiftLintBuiltInRules/Models/BuiltInRules.swift
@@ -149,6 +149,7 @@ public let builtInRules: [any Rule.Type] = [
     OverrideInExtensionRule.self,
     PatternMatchingKeywordsRule.self,
     PeriodSpacingRule.self,
+    PreferAssetSymbolsRule.self,
     PreferConditionListRule.self,
     PreferKeyPathRule.self,
     PreferNimbleRule.self,